+++ CHANGELOG.md
@@ -21,7 +21,8 @@
 ## [Unreleased]
 
 ### Добавлено
-
+- Реализован провайдер `S3Provider` для работы с S3 (методы загрузки, получения, удаления и листинга файлов);
+- Написаны юнит-тесты для `S3Provider` с использованием pytest и pytest-asyncio, включая мокирование aioboto3 и асинхронную пагинацию;
 - Конфигурация линтеров flake8, black, isort и mypy в pyproject.toml;
 - flake8 настроен с inline-quotes = "'", max-line-length = 79, docstring-convention = "google", require-annotations = true, max-complexity = 5, игнорированием ошибки ANN101,D104;
 - black с длиной строки 79 и отключённой нормализацией кавычек;